% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call_Data.R
\name{call.data}
\alias{call.data}
\title{call.data}
\usage{
call.data(id.indicator, language.en = TRUE)
}
\arguments{
\item{id.indicator}{It requests the number of the indicator according to the names consulted in the call.indicators function.}

\item{language.en}{If language.en is TRUE, it will return the names of the indicators in English. If it is FALSE, it will return the names of the indicators in Spanish.}
}
\description{
You can consult the data of CEPALSTAT indicators.
}
