% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call_Indicators.R
\name{call.indicators}
\alias{call.indicators}
\title{You can consult the names of the CEPALSTAT indicators}
\usage{
call.indicators(language.en = TRUE)
}
\arguments{
\item{language.en}{If language.en is TRUE, it will return the names of the indicators in English. If it is FALSE, it will return the names of the indicators in Spanish.}
}
\description{
You can consult the names of the CEPALSTAT indicators
}
